refactor(counters_stable): move page object
