+++ examples/counters_stable/e2e/tests/add_counter.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { CountersPage } from "./counters_page";
+import { CountersPage } from "./fixtures/counters_page";
 
 test.describe("Add Counter", () => {
   test("should increase the number of counters", async ({ page }) => {
