+++ examples/counters_stable/e2e/tests/fixtures/counters_page.ts
