+++ examples/counters_stable/e2e/tests/view_counters.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { CountersPage } from "./counters_page";
+import { CountersPage } from "./fixtures/counters_page";
 
 test.describe("View Counters", () => {
   test("should see the title", async ({ page }) => {
